+++ benchmark/ssx.jsx
@@ -1,0 +1,9 @@
+export default function (lenght = 1000) {
+  return (
+    <>
+      {new Array(lenght).fill(null).map((v, key) => (
+        <b key={key} class="foo">bar</b>
+      ))}
+    </>
+  );
+}