+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [0.1.4] - Unreleased
+### Added
+- Benchmark to compare with React and Preact.
+
+### Changed
+- Make `ssxElement` property not enumerable.
+
 ## [0.1.3] - 2025-01-06
 ### Added
 - Export `renderComponent` function.
@@ -30,6 +37,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ## [0.1.0] - 2024-07-20
 First version
 
+[0.1.4]: https://github.com/oscarotero/ssx/compare/v0.1.3...HEAD
 [0.1.3]: https://github.com/oscarotero/ssx/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/oscarotero/ssx/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/oscarotero/ssx/compare/v0.1.0...v0.1.1